LayoutTests/imported/w3c/web-platform-tests/cookies/samesite/about-blank-subresource.https-expected.txt


FAIL SameSite cookies on subresource of top-level about:blank window assert_equals: `samesite_lax=0.05247175901404755` in request to `https://localhost:9443`. expected (string) "0.05247175901404755" but got (undefined) undefined
